requirements.txt
#These are the requirements you need to run this project\
## run the command 'pip install -r requirements.txt'
streamlit==1.22.0    #these are all boiler plate stuff for now
flask==2.2.5           
PyPDF2==3.0.1
scikit-learn==1.3.0
nltk==3.8.1
mysql-connector-python==8.0.33
pymysql>=1.0.3
python-dotenv>=1.0.0
